Merges CDFileClass to common based on RA version.
